CHUNK_SIZE is u32

+++ src/multiexp.rs
@@ -374,7 +374,7 @@ pub fn gpu_multiexp_consistency() {
     const CHUNK_SIZE: usize = 1048576;
     const MAX_LOG_D: usize = 20;
     const START_LOG_D: usize = 10;
-    let mut kern = gpu::MultiexpKernel::<Bls12>::create(CHUNK_SIZE).ok();
+    let mut kern = gpu::MultiexpKernel::<Bls12>::create(CHUNK_SIZE as u32).ok();
     if kern.is_none() { panic!("Cannot initialize kernel!"); }
     let pool = Worker::new();
